Support multipart upload when image_size == 0
When doing an image upload of a volume from cinder-volume to glance with
an s3 backend, image_size in this context is zero. As a result the check
that image_size is larger than s3_store_large_object_size is always
negative and glance_store never uses multipart upload.

As a workaround, allow setting s3_store_large_object_size to 0, which
forces glance_store to use multipart upload.

Related-Bug: #2043740
Change-Id: I436d56d0489b081822ed55108e10d43cc38512bb